requirements.txt
newspaper3k>=0.2.8
openai>=1.8.0
